welcome: show request token btn when appropriate
closes #93
