+++ reana-ui/src/pages/workflowList/components/Welcome.module.scss
@@ -14,8 +14,4 @@
   padding: 1em;
   margin: 4em 0;
   color: darken($gray, 10%);
-
-  :global(.ui.header).header {
-    margin-bottom: 1em;
-  }
 }
